add image wrong, but everything is good

+++ backend/db/migrations/20231222205011-create-spot.js
@@ -42,11 +42,11 @@ module.exports = {
         allowNull:false
       },
       lat: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL,
         allowNull:false
       },
       lng: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL,
         allowNull:false
       },
       name: {
@@ -58,7 +58,7 @@ module.exports = {
         allowNull:false
       },
       price: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL,
         allowNull:false
       },
       createdAt: {
